jest: Update all Jest-related dependencies to their latest.
In particular, this gets us Jest 26.4.1, which contains
facebook/jest#10414. This should fix an issue [1] we'd otherwise
encounter very soon, where differences in a URL object encountered
with `expect(...).toBe(...)` or `expect(...).toEqual(...)` would
result in a very cryptic error message instead of a straightforward
test failure with the desired explanation.

This upgrade was pretty easy to do! I'm hoping fb23341c3 has
increased the ease of staying up-to-date with Jest.

Since we're going from a 22.x to 23.x of eslint-plugin-jest, we
check the declared breaking changes at 23.0.0 [2]. The only one that
concerns us is that part of the job of `jest/valid-describe` has
been transferred to `jest/valid-title`, and `jest/valid-title` has
been added to the `jest/recommended` config, which we take. The
validation in question is silly, and unfortunately, we can't
reasonably take all of `jest/valid-title` except that part, so, turn
it all off.

See also some prep work for this upgrade in a recent commit.

[1] https://github.com/facebook/jest/issues/10408
[2] https://github.com/jest-community/eslint-plugin-jest/releases/tag/v23.0.0