deps: Add `typescript` as dev dependency.
We don't need TypeScript ourselves [1], but we want to upgrade
`eslint-plugin-jest` soon, and a rather distant dependency has a
peer dependency of TypeScript (as we also saw in 01593b3af). Here's
the output:

```
warning "eslint-plugin-jest > @typescript-eslint/experimental-utils
> @typescript-eslint/typescript-estree > tsutils@3.17.1" has unmet
peer dependency "typescript@>=2.8.0 || >= 3.2.0-dev || >= 3.3.0-dev
|| >= 3.4.0-dev || >= 3.5.0-dev || >= 3.6.0-dev || >= 3.6.0-beta ||
>= 3.7.0-dev || >= 3.7.0-beta".
```

Ah well, at least we can put it in `devDependencies`. And
`yarn why typescript` reveals that `prettier-eslint` and
`prettier-eslint-cli` have both already been bringing it in as their
dependency. Those libraries are using version 3.9.7, so we set the
range to include that version, so we can avoid pulling in a new
copy.

[1] #3458 is open to consider migrating from Flow to TypeScript, but
    that's not what this commit is about.