android build: Prevent build failure with duplicate libc++_shared.so.
Part of the RN v0.61 -> v0.62 changes to the template app [1],
corresponding to facebook/react-native@2fd50882b. It's described as
a way to handle the case where dependencies, such as Flipper (a
thing newly available with the upgrade), "might" also depend on
libc++_shared.so, causing build failures.

So, it sounds like it belongs before the commits that add Flipper
dependencies. I didn't get any build failures or runtime errors; if
I had, it might have been a sign that this must go at or after the
main upgrade commit.

[1] https://react-native-community.github.io/upgrade-helper/?from=0.61.5&to=0.62.2