common: Delete unused `Arc` and `Circle`.
We stopped using these in 9fb077277 when we landed on a better
solution. From a brief skim of the RN issue indicated in that
commit, it seems like it may have been fixed in RN v0.61. But we're
happy with our PNG-based implementation of SpinningProgress, at
least for now.

In any case, these components depend on something called `ART`,
which used to come from RN core but, in RN v0.62, was moved to
`react-native-community` as part of the lean-core effort. No need to
add that dependency if we're not going to use it.

See also discussion at
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/RN's.20.22ART.22.20API/near/1004888.