compose: Don't show video button for unsupported providers.
Clients that connect to a server that has configured a video call
provider not currently supported by the mobile app should not display a
video call button in the compose menu.

This change refactors the realm slice of the redux state, removing the
jitsiServerUrl field and its associated selector in favor of
encapsulating that data in a disjoint union that contains members for
each supported video chat provider and its associated data.
Additionally, it updates `src/compose/ComposeBox.js` to conditionally
create the video link insertion handler based on the current provider.

One issue I came across when implementing this code is the type of the
videoChatProvider prop in ComposeBox, which Flow is erroneously widening
to also allow generic strings as part of the union.

Fixes: #4246