flipper: Enable Flipper.
Call the appropriate initialize-Flipper function in native runtime
code.

Part of the RN v0.61 -> v0.62 changes to the template app [1],
corresponding to facebook/react-native@05f5cb534. This should be
fine to do after the main upgrade commit; nothing in React Native's
internals should depend on Flipper being enabled [2].

On iOS, use the form of the initialize-Flipper function as it was
updated in facebook/react-native@b4d1fcfb2 and wrap the call site in
a conditional as done there.

[1] https://react-native-community.github.io/upgrade-helper/?from=0.61.5&to=0.62.2
[2] https://github.com/zulip/zulip-mobile/pull/4244#discussion_r482536188