nav: Enable swiping between top tabs.
In one of the recent React Navigation upgrades, a rapid sideways
scroll animation was added for when you tap another tab to switch to
it. This feels better than having no animation, but it really
suggests that swiping ought to work to switch between tabs [1]. We
haven't been offering that; so, do that.

[1] https://github.com/zulip/zulip-mobile/pull/4249#issuecomment-691388484

Fixes: #3997