deps: Upgrade `react-native-unimodules` to ^0.10.1, the latest.
Also change the `@unimodules/core` version to match, as we did in
e116c3f9d.

Nothing in my reading suggests that we can't do this before the RN
v0.61 -> v0.62 upgrade, but I figured the chance of success would be
greater if we did it after the upgrade.

No breaking changes were specifically announced in the
`react-native-unimodules` changelog [1]; however, it does say,
"Updated dependencies to match versions included in Expo SDK38." To
be safe, I checked the list of breaking changes in the release notes
for Expo SDK 38 [2]. Nothing there seems to apply to us; there are
several announcements about breaking changes in `expo-*` packages,
but we don't use any in that list.

Like in e116c3f9d, we see several changes to `ios/Podfile.lock`.
Unlike in e116c3f9d, we don't see any changes on the Android side,
even after building and running the app.

After this update, I didn't encounter any build failures or runtime
errors on iOS or Android.

[1] https://github.com/expo/expo/blob/master/packages/react-native-unimodules/CHANGELOG.md#0101--2020-05-29
[2] https://github.com/expo/expo/blob/master/CHANGELOG.md#3800