webview: Prepare to "revive" realm in `handleInitialLoad`.
In an upcoming commit, we'll change the `Auth` type to have a URL
object for the realm. We call `JSON.stringify` on the arguments
passed to `handleInitialLoad`. For the `realm` property, this means
it'll be a string URL, not a URL object. Here, we pick it out in
preparation for "reviving" it back into its URL form in the body of
`handleInitialLoad`.