apiFetch: Use URL computation in `apiFetch`.
For #4146, but also to fix a recent regression from 865914f0e.

The URL constructor already throws a special "Invalid URL" error if
it can't put together a valid URL from the inputs, so, nicely, we
get to remove the `isValidUrl` call site.

There was a double-forward-slash issue in this string concatenation:
the realm gained a trailing slash with the round-trip through
`new URL(…).toString()` from 865914f0e, and we kept the leading slash
in what we were adding to the end of the realm.