navSelectors: Dissolve `getCanGoBack`.
Working toward #3804.

We're about to reimplement most of the functions in navSelectors.js
so that they grab navigation state from the recently added
`NavigationService` instead of from `state.nav` in Redux.

For now, it's convenient to keep sourcing data from `state.nav` in
just a few places. In this case, the one call site of `getCanGoBack`
(the `connect` wrapper for `BackNavigationHandler`) is placed such
that it's not convenient for it to get navigation state from
`NavigationService`. In particular, `NavigationService` relies on a
ref being set -- to a component that's deeper down in the component
hierarchy. If we ask `NavigationService` for the navigation state,
that component won't have been constructed, the ref will come up
empty, and we'll have an error.

`BackNavigationHandler` can disappear as #3804 is completed; you can
see how and why this is done in an upcoming commit.