i18n: Cut another string we don't actually use, "Narrow".
Noticed this at the same time as "Narrow to conversation",
discussed in the previous commit.

This one was added recently to the message list for translation,
in 17f73d88d.  But as far as I can tell, we never actually added
any UI that used this string, either in that commit or any since.
I also searched (`git grep -w Narrow src/`) for any existing
reference, and though it's a bit noisy because the string is short
and coincides with a type name we use a lot, I didn't find any
string uses that way either.