navReducer: Use `getNavigationRoutes`.
Also add a comment, with Greg's description [1] of what's going on here.

We have this handy function for grabbing what we want from the
navigation state, so we might as well use it. When we use this
function, the navigation state is gotten from `NavigationService`,
not from Redux, which moves us closer to #3804.

To be more rigorous, though, we wouldn't want this to stay this way
in `navReduce` forever: reducers are supposed to be pure functions
of state and actions. But the incorrectness isn't obviously harmful
here, and we're about to transplant this code somewhere else, and
it's helpful to separate a small tweak like this from the commit
where we move the code.

[1] https://github.com/zulip/zulip-mobile/pull/4274#discussion_r500688470