android: Update target SDK version to 29, i.e. Android 10.
This change will be required in order to upload new releases to the
Play Store, effective 2020-11-01, a couple of weeks from now.

The change causes Android 10 and later to apply to our app a number
of behavior changes introduced in that version, documented here:
  https://developer.android.com/about/versions/10/behavior-changes-10

After reading carefully through that documentation, I believe we
can rule out almost all the changes as possibly affecting us:
  https://github.com/zulip/zulip-mobile/issues/3665#issuecomment-710720020
except for one -- "scoped storage" -- and it seems likely that the
remaining don't either.  We've opted out of scoped storage for now,
in the previous commit.

We'll want to thoroughly test a beta version with this change
before shipping it broadly, to try to catch any remaining issues.

Fixes: #3665