initialDataTypes: Add `RawInitialData`.
Greg points out [1] that it's time we start keeping track of a type
for our initial data pre-transformation, since we're about to start
transforming it.

Here, we say what we expect `realm_users` and two of its friends to
look like. We grab those expectations from the `avatar_url` field on
`User` and on `CrossRealmBot`.

[1] https://github.com/zulip/zulip-mobile/pull/4230#discussion_r513100837