narrow: Fix bug where self-1:1 "is in" every PM conversation.
We had a hilarious bug: if you've already visited some group PM
conversation (so that we're following it in `state.narrows`), and
then you send yourself a 1:1 PM -- either from some other client,
or from this same client -- then the 1:1 will appear to show up in
that group conversation.

This doesn't affect what messages exist on the server, or what
messages anyone else sees.  And it can only happen with a message you
yourself sent, because it's limited to self-PMs.  Still -- it could
potentially *look* to you like something quite awkward had happened!
Which would, itself, be awkward.

The root of the bug is a funny `normalizedRecipients === ownEmail`
case in the `isMessageInNarrow` function.  That, in turn, seems to be
there (it was added in d8faae9d7, which introduced this bug) to try to
correct for where we add `ownEmail` to the list of emails on the line
before... which in turn is to try to correct for how in many places we
leave out the self-user from describing certain PM narrows.

We can fix the bug by consistently using that latter normalization,
quirky though it is, within that function.  The email list
representing a group PM narrow should already be normalized in that
way, leaving out the self user; and the old code here was already
relying on that.

Fixes: #3324