UserAvatarWithPresence: Stop double-processing avatar URLs.
Before this commit, `getAvatarFromMessage`, which is a thin wrapper
on `getAvatarUrl`, had a callsite in `Lightbox`; the output of that
call then made its way to `UserAvatarWithPresence`, and the latter
passed it to `getAvatarUrl` again.

This situation of passing the `.avatar_url` field on a message,
user, or cross-realm bot through `getAvatarUrl` twice, instead of
just once, has so far been harmless; the second time has just
returned its input unmodified. But if we were to do the same thing
with a `size` over 100, we could end up with something like
`.../foo-medium-medium.png`, for an uploaded avatar, which would be
no good.

So, stop the double-processing by removing the
`getAvatarFromMessage` call. And try to prevent the same thing from
happening again by changing the contract of `UserAvatarWithPresence`
(since it calls `getAvatarUrl`), so that it clearly asks for the
`.avatar_url` property on a message, user, or cross-realm bot
object, and propagate the change of contract to a few callers,
adding and changing some JSDocs.

The change of contract is safe because `getAvatarUrl` knows how to
handle the `.avatar_url` field on message, user, and cross-realm bot
objects.