webview: Fix a "jumpy animation glitch on spoiler collapse".
Following zulip/zulip@f3011d3b7.

The particular glitch (not explained in that commit) is that, before
the animation to shrink the spoiler content begins, the height is
instantaneously set to be bigger than it should be. The intent is to
set the explicit 'height' attribute to match the computed content
height, but it was actually getting set to a value (.scrollHeight)
that includes vertical padding.

See discussion at
https://github.com/zulip/zulip-mobile/pull/4169/files/c0ee9183d56417f878565b384261b245ecbb3ce6#r522357362.