events: Drop "batching" of actions.
It's extremely rare that this has an opportunity to happen; so even if
there's some optimization here in principle, we're not getting any
significant value from it.  And it just makes the control flow more
complicated.

Empirically: I happened to run the app for an hour today, logged into
chat.zulip.org as a test user, and logging all Redux actions.  This
batching happened just once.

More theoretically: The point of this library is to batch up some
Redux actions so that subscribers get notified just once for the whole
batch -- in particular, so `connect`ed React components don't rerender
repeatedly.  At best, this is helpful only when we have two actions
happening in the timespan of a single frame (which is 16.7ms on most
devices, and can be shorter on recent high-end devices).  And it's
just going to be extremely uncommon to have more than one Zulip event
from the server in a span of 17ms.  If that were routine, it'd cause
scaling problems elsewhere first.

I'm also not sure this batching is ever necessary in the first place
-- if it is, it's fundamentally a workaround for an optimization
missing in React itself and/or react-redux.  I'm pretty sure I've seen
release notes for both those libraries discuss getting better at this
sort of optimization; so if there is still something to work around,
there likely won't be before long.