AccountDetails: Fix slight misalignment of status dot.
In the full-screen profile, the status dot's vertical alignment with
the user's name has been just a bit off. It's also been placed to
the right of the name, instead of the left. In the webapp's right
sidebar, it goes on the left of the name; might as well match that.

So, fix the vertical alignment and put the dot on the left of the
name.

The vertical-alignment story is a little tricky [1]. The name and
the status dot have already both been vertically centered in their
parent. (That parent, the `View` with
`componentStyles.statusWrapper`, has `flexDirection: 'row'`, and it
has `alignItems` set to 'center'. The `alignItems` comes rather
indirectly from `componentStyles.componentListItem`, through
`ComponentList`.) Still, the text and the image still seem
misaligned -- so we go with the plausible solution of nudging the
status dot down by a logical pixel or two.

Co-authored-by: Chris Bobbe <cbobbe@zulip.com>
Fixes: #4307

[1] https://github.com/zulip/zulip-mobile/pull/4316#pullrequestreview-537680616