IosCompliantAppleAuthButton types: Limit what can get passed for `style`.
Partly in the spirit of 8b30fa9d6, but also because this helps us
avoid lots of mystifying error reports from Flow when we take the
upgrade to Flow v0.122.0, with the React Native v0.63 upgrade
(#4245).

See discussion of those errors [1], and Greg's conclusion that Flow
might have been a bit more helpful than it was here. Ah, well.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Flow.20errors.20in.20.23M4318/near/1075420