deps: Upgrade ESLint-related deps to their latest.
Done with

```
yarn upgrade --latest $(tools/test deps | grep eslint)
```

which was easier than 01593b3af.

One thing we neglected to mention in 01593b3af, which still applies
here, is that the version of ESLint used by `prettier-eslint-cli` is
still 5, and it's unclear when (or if) that'll change [1], which is
unfortunate.

[1] https://github.com/prettier/prettier-eslint-cli/issues/304

Fixes: #4253