docs/style: Tighten style of new style-guide entry.
Generally I feel like our style guide runs long; so I've tried to
have each entry get right to the point of what to do or not do, with
minimal rationale and a link to more elsewhere when appropriate.

Also add a few words about concretely how to find what names to avoid.