android build: Upgrade Gradle wrapper to match Gradle, now and in future.
The way we've historically managed this -- using the upgrade command
in the most natural way -- has meant that we're always using the
Gradle wrapper from the *previous* Gradle version we last upgraded from.

The version mismatch itself is fine: much of the point of the Gradle
wrapper is to be a thin indirection layer that can start up whatever
version of Gradle a given project specifies.  But when doing an
upgrade, this behavior sure makes things confusing -- the shell script
we end up with can vary when trying two different sequences of
upgrades that end with the exact same version.  That sure looks like
something's wrong with the upgrade process, causing time spent debugging.

Now that we have debugged it, there's a simple workaround to make it
stop happening: just run each upgrade twice, so that everything settles
at the new version.  Do that for the last upgrade, and update the comment
so we know to do that in the future.