narrow [nfc]: Add comments cross-linking getNarrowsForMessage and its converse.
Whenever we add a new type of narrow, this function will probably need
to gain a bit of logic to include that type of narrow where appropriate.

There's no obvious way to get the type-checker to remind us to do
that, because here we're conditionally building up a Narrow value
rather than breaking one down.  But the closely-related function
isMessageInNarrow does have an actual Narrow object to break down, so
it enjoys the benefit of Flow confirming it's covered all the cases.
Add a comment there pointing here, in hopes of riding on that.

Also add a comment about SEARCH_NARROW so that this function's
implementation explicitly mentions every type of narrow; and
add cross-references to the jsdoc.